requirements.txt
appdirs==1.4.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coloredlogs==15.0
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
humanfriendly==9.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psutil==5.8.0
psycopg2-binary
requests==2.25.1
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
demjson==2.2.4